BHttpRequest: Fix the cast and therefor GCC2 build
Reviewed-on: https://review.haiku-os.org/c/haiku/+/3614
(cherry picked from commit 8c812f2d6380f339d01b32a547424619b3f75408)

+++ src/kits/network/libnetapi/HttpRequest.cpp
@@ -1207,5 +1207,5 @@ BHttpRequest::_NotifyDataReceived(const char* data, off_t pos, ssize_t size,
 	}
 	fListener->DataReceived(this, data, pos, size);
 	fListener->DownloadProgress(this, bytesReceived,
-		std::max((off_t)0, bytesTotal));
+		std::max((ssize_t)0, bytesTotal));
 }
